requirements.txt
kombu==4.2.1
celery==4.2.1
flower==0.9.2
requests
pymongo==3.7.2
flower==0.9.2
python-dateutil==2.7.5
pytz
pyyaml
python-json-logger==0.1.10
Jinja2==2.10
tornado>=4.2.0,<6.0.0
